# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Profo 房产数据中心",
    "description": "轻量级、本地化、高性能的房产数据仓库系统",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/projects/{project_id}/selling/roles": {
      "put": {
        "tags": [
          "projects",
          "projects",
          "sales"
        ],
        "summary": "Update Sales Roles",
        "description": "更新销售角色 (Sync)",
        "operationId": "update_sales_roles_api_v1_projects__project_id__selling_roles_put",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "项目ID",
              "title": "Project Id"
            },
            "description": "项目ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SalesRolesUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/selling/viewings": {
      "post": {
        "tags": [
          "projects",
          "projects",
          "sales"
        ],
        "summary": "Create Viewing Record",
        "description": "创建带看记录 (Sync)",
        "operationId": "create_viewing_record_api_v1_projects__project_id__selling_viewings_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "项目ID",
              "title": "Project Id"
            },
            "description": "项目ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SalesRecordCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SalesRecordResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/selling/offers": {
      "post": {
        "tags": [
          "projects",
          "projects",
          "sales"
        ],
        "summary": "Create Offer Record",
        "description": "创建出价记录 (Sync)",
        "operationId": "create_offer_record_api_v1_projects__project_id__selling_offers_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "项目ID",
              "title": "Project Id"
            },
            "description": "项目ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SalesRecordCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SalesRecordResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/selling/negotiations": {
      "post": {
        "tags": [
          "projects",
          "projects",
          "sales"
        ],
        "summary": "Create Negotiation Record",
        "description": "创建面谈记录 (Sync)",
        "operationId": "create_negotiation_record_api_v1_projects__project_id__selling_negotiations_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "项目ID",
              "title": "Project Id"
            },
            "description": "项目ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SalesRecordCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SalesRecordResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/selling/records": {
      "get": {
        "tags": [
          "projects",
          "projects",
          "sales"
        ],
        "summary": "Get Sales Records",
        "description": "获取销售记录 (Sync)",
        "operationId": "get_sales_records_api_v1_projects__project_id__selling_records_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "项目ID",
              "title": "Project Id"
            },
            "description": "项目ID"
          },
          {
            "name": "record_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "记录类型筛选",
              "title": "Record Type"
            },
            "description": "记录类型筛选"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  },
                  "title": "Response Get Sales Records Api V1 Projects  Project Id  Selling Records Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/selling/records/{record_id}": {
      "delete": {
        "tags": [
          "projects",
          "projects",
          "sales"
        ],
        "summary": "Delete Sales Record",
        "description": "删除销售记录 (Sync)",
        "operationId": "delete_sales_record_api_v1_projects__project_id__selling_records__record_id__delete",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "项目ID",
              "title": "Project Id"
            },
            "description": "项目ID"
          },
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "记录ID",
              "title": "Record Id"
            },
            "description": "记录ID"
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProjectResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "项目ID"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name",
            "description": "项目名称"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "项目状态"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "community_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Community Name"
          },
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "area": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Area"
          },
          "layout": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Layout"
          },
          "orientation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Orientation"
          },
          "is_deleted": {
            "type": "boolean",
            "title": "Is Deleted",
            "default": false
          },
          "renovation_stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Renovation Stage"
          },
          "contract_no": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contract No",
            "description": "合同编号"
          },
          "signing_price": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signing Price",
            "description": "签约价格(万)"
          },
          "signing_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signing Date"
          },
          "signing_period": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signing Period"
          },
          "extension_period": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extension Period"
          },
          "extension_rent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extension Rent"
          },
          "cost_assumption": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost Assumption"
          },
          "planned_handover_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Planned Handover Date"
          },
          "other_agreements": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Other Agreements"
          },
          "contract_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contract Status"
          },
          "owner_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Name"
          },
          "owner_phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Phone"
          },
          "owner_id_card": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Id Card"
          },
          "owner_info": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Info"
          },
          "list_price": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "List Price",
            "description": "挂牌价(万)"
          },
          "listing_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Listing Date"
          },
          "sold_price": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sold Price"
          },
          "sold_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sold Date"
          },
          "transaction_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transaction Status"
          },
          "total_income": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Income",
            "default": "0"
          },
          "total_expense": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Expense",
            "default": "0"
          },
          "net_cash_flow": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Net Cash Flow",
            "default": "0"
          },
          "roi": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Roi",
            "default": 0.0
          },
          "signing_materials": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/SigningMaterial"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signing Materials",
            "description": "签约材料列表"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "ProjectResponse",
        "description": "项目完整响应模型 - 适配新的规范化表结构\n业务日期字段使用字符串类型 (YYYY-MM-DD) 避免时区问题"
      },
...
      "RecordType": {
        "type": "string",
        "enum": [
          "viewing",
          "offer",
          "negotiation"
        ],
        "title": "RecordType",
        "description": "销售记录类型枚举"
      },
...
      "SalesRecordCreate": {
        "properties": {
          "record_type": {
            "$ref": "#/components/schemas/RecordType"
          },
          "customer_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Name"
          },
          "customer_phone": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Phone"
          },
          "customer_info": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Info"
          },
          "record_date": {
            "type": "string",
            "format": "date-time",
            "title": "Record Date"
          },
          "record_time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Record Time"
          },
          "price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "feedback": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feedback"
          },
          "result": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "related_agent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Related Agent"
          }
        },
        "type": "object",
        "required": [
          "record_type",
          "record_date"
        ],
        "title": "SalesRecordCreate",
        "description": "创建销售记录"
      },
      "SalesRecordResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "record_type": {
            "type": "string",
            "title": "Record Type"
          },
          "customer_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Name"
          },
          "customer_phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Phone"
          },
          "customer_info": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Info"
          },
          "record_date": {
            "type": "string",
            "format": "date-time",
            "title": "Record Date"
          },
          "record_time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Record Time"
          },
          "price": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "feedback": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feedback"
          },
          "result": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "related_agent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Related Agent"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "record_type",
          "customer_name",
          "customer_phone",
          "customer_info",
          "record_date",
          "record_time",
          "price",
          "notes",
          "feedback",
          "result",
          "related_agent",
          "created_at"
        ],
        "title": "SalesRecordResponse",
        "description": "销售记录响应"
      },
      "SalesRolesUpdate": {
        "properties": {
          "channel_manager": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Channel Manager"
          },
          "presenter": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Presenter"
          },
          "negotiator": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Negotiator"
          },
          "property_agent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Property Agent"
          },
          "client_agent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Agent"
          },
          "first_viewer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Viewer"
          }
        },
        "type": "object",
        "title": "SalesRolesUpdate",
        "description": "更新销售角色"
      },
      "SigningMaterial": {
        "properties": {
          "filename": {
            "type": "string",
            "title": "Filename",
            "description": "文件名"
          },
          "url": {
            "type": "string",
            "title": "Url",
            "description": "文件URL"
          },
          "category": {
            "type": "string",
            "title": "Category",
            "description": "附件分类"
          },
          "fileType": {
            "type": "string",
            "title": "Filetype",
            "description": "文件类型"
          },
          "size": {
            "type": "integer",
            "title": "Size",
            "description": "文件大小(字节)",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "filename",
          "url",
          "category",
          "fileType"
        ],
        "title": "SigningMaterial",
        "description": "签约材料附件"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/token"
          }
        }
      }
    }
  }
}